Refactor webapp into modules
